+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Breaking Changes
+
+- `CreatorInterface::create`'s parameter `$constructor` is now of type `string`
+  instead of `string|null`. 
+
+
 [0.3.0](https://github.com/coniadev/wire/releases/tag/0.3.0) - 2024-01-18
 -------------------------------------------------------------------------